+++ x/session/expected/keeper.go
@@ -19,20 +19,20 @@ type DepositKeeper interface {
 
 type PlanKeeper interface {
 	HasNodeForPlan(ctx sdk.Context, id uint64, address hubtypes.NodeAddress) bool
-	GetNodesForPlan(ctx sdk.Context, id uint64, skip, limit int) node.Nodes
+	GetNodesForPlan(ctx sdk.Context, id uint64, skip, limit int64) node.Nodes
 }
 
 type NodeKeeper interface {
-	GetNodes(ctx sdk.Context, skip, limit int) node.Nodes
+	GetNodes(ctx sdk.Context, skip, limit int64) node.Nodes
 }
 
 type SubscriptionKeeper interface {
 	GetSubscription(ctx sdk.Context, id uint64) (subscription.Subscription, bool)
 
-	GetSubscriptionsForNode(ctx sdk.Context, address hubtypes.NodeAddress, skip, limit int) subscription.Subscriptions
+	GetSubscriptionsForNode(ctx sdk.Context, address hubtypes.NodeAddress, skip, limit int64) subscription.Subscriptions
 	HasSubscriptionForNode(ctx sdk.Context, address hubtypes.NodeAddress, id uint64) bool
 
-	GetActiveSubscriptionsForAddress(ctx sdk.Context, address sdk.AccAddress, skip, limit int) subscription.Subscriptions
+	GetActiveSubscriptionsForAddress(ctx sdk.Context, address sdk.AccAddress, skip, limit int64) subscription.Subscriptions
 
 	SetQuota(ctx sdk.Context, id uint64, quota subscription.Quota)
 	GetQuota(ctx sdk.Context, id uint64, address sdk.AccAddress) (subscription.Quota, bool)
